+++ NEWS.md
@@ -1,3 +1,5 @@
+# DO.utils (development version)
+
 # DO.utils 0.2.7
 
 ## Dependencies